server/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
requests==2.23.0
Pillow==7.1.2

# search
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.1
Whoosh==2.7.4
beautifulsoup4==4.9.0
gunicorn==20.0.4
itsdangerous==1.1.0
lxml==4.5.0
cheshire3==1.1.8
docutils==0.16
